Update of base image and reproducibility testing
Workshops hasn't been modified in a little while and we were having some
trouble reproducing the production build. This commit is part of a
series investigating the build process and trying to make sure that it
is completely reproducible. Right now there are some build steps which
happen at runtime which don't seem to be 100% deterministic and rely on
external volumes. We'd like to make sure those are understood and made
as reproducible as possible.

I'm not sure what has happened to the main/master branch of this repo.
It exists on Garibaldi but not on GitHub. Fortunately, the head of
master on Garibaldi is a direct ancestor of the GitHub development
branch, so we have just created a new branch (wtf) from that commit
(04d418b6). This branch is not meant to be permanent but rather to allow
us to see what we want to change. The baseline test for this branch will
be to be able to restore from a production DB dump and verify workshops
functionality. Ultimately though, we would like to reinstate a stable and
reproducible main branch which incorporates the lessons learned in this
branch and is integrated with the CI system.

This commit updates the base image to passenger-ruby27:2.3.0 and there
have been extensive modifications to both the Dockerfile and the
entrypoint script. Ruby is bumped from 2.7.4 to 2.7.6 and node is
updated to node 16.x. There have also been some minor modifications to
the docker-compose.yml file.

Signed-off-by: Ian Allison <iana@pims.math.ca>