README and minor package bump
Removing an unused instruction from the restore section of the README
and tagging in the latest yarn.lock which agrees with our package
selection. The entrypoint script was running an yarn update which I've
removed for now and this commit preserves the latest version of the
lockfile generated by that `yarn update`.

Signed-off-by: Ian Allison <iana@pims.math.ca>